+++ templates/server-shelf/bin/server.dart
@@ -4,26 +4,29 @@ import 'package:args/args.dart';
 import 'package:shelf/shelf.dart' as shelf;
 import 'package:shelf/shelf_io.dart' as io;
 
-void main(List<String> args) {
+main(List<String> args) async {
   var parser = new ArgParser()
     ..addOption('port', abbr: 'p', defaultsTo: '8080');
 
   var result = parser.parse(args);
 
-  var port = int.parse(result['port'], onError: (val) {
-    stdout.writeln('Could not parse port value "$val" into a number.');
-    exit(1);
-  });
+  var port = int.tryParse(result['port']);
+
+  if (port == null) {
+    stdout.writeln(
+        'Could not parse port value "${result['port']}" into a number.');
+    // 64: command line usage error
+    exitCode = 64;
+    return;
+  }
 
   var handler = const shelf.Pipeline()
       .addMiddleware(shelf.logRequests())
       .addHandler(_echoRequest);
 
-  io.serve(handler, '0.0.0.0', port).then((server) {
-    print('Serving at http://${server.address.host}:${server.port}');
-  });
+  var server = await io.serve(handler, 'localhost', port);
+  print('Serving at http://${server.address.host}:${server.port}');
 }
 
-shelf.Response _echoRequest(shelf.Request request) {
-  return new shelf.Response.ok('Request for "${request.url}"');
-}
+shelf.Response _echoRequest(shelf.Request request) =>
+    new shelf.Response.ok('Request for "${request.url}"');